- Fix typings - Move TruffleContract to typings
