+++ packages/configs/tslints/base-tslint.json
@@ -7,6 +7,7 @@
   ],
   "rules": {
     "prettier": true,
+    "no-implicit-dependencies": [true, "dev"],
     "interface-name": false,
     "object-literal-sort-keys": false,
     "max-classes-per-file": false,
