+++ packages/typescript-typings/tsconfig.json
@@ -1,7 +1,7 @@
 {
-    "extends": "../../tsconfig",
-    "compilerOptions": {
-        "outDir": "lib"
-    },
-    "include": []
+  "extends": "./node_modules/@counterfactual/configs/tsconfigs/base-tsconfig",
+  "compilerOptions": {
+    "outDir": "lib"
+  },
+  "include": []
 }
