+++ packages/typescript-typings/types/counterfactual/index.d.ts
@@ -1,0 +1,6 @@
+export interface TruffleContract {
+  readonly contractName: string;
+  readonly abi: any[];
+  readonly bytecode: string;
+  readonly networks: { [networkName: string]: { address: string } };
+}
