+++ packages/typescript-typings/types/index.d.ts
@@ -1,0 +1,5 @@
+export * from "./chai";
+export * from "./chai-bignumber";
+export * from "./counterfactual";
+export * from "./solc";
+export * from "./truffle";
